First working linked list
This is the implementation of a stack (backed by a linked list) that is
translated from [Too many linked
lists](https://rust-unofficial.github.io/too-many-lists/second.html)
book.
In this implementation, no capture checking and linear type feature is
used. It means that the borrowing mechanism is not working at all.

Also, it has only _should work_ test, not _should not work_ tests.