Implement stacked borrowed model
`InternalRef` implements stacked borrowed model, and the linked list
implementation and _should work_ tests are changed to use the API.

There are couple of points in the implementation that are marked as
*TODO*s which are signs that the current system is exploitable. Some of
them should be solved through static restrictions, and some of them
maybe through runtime. By this commit _should not work_ tests are not
refactored.